pull updateblob pieces into separate go files
