+++ pkg/cluster/deploy.go
@@ -1,13 +1,11 @@
 package cluster
 
 import (
-	"bytes"
 	"context"
 	"crypto/sha256"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
 
 	"github.com/openshift/openshift-azure/pkg/api"
 	"github.com/openshift/openshift-azure/pkg/jsonpath"
@@ -46,15 +44,9 @@ func (u *simpleUpgrader) Deploy(ctx context.Context, cs *api.OpenShiftManagedClu
 	return nil
 }
 
-type updateblob map[instanceName]hash
 type scalesetName string
 type instanceName string
 type hash string
-
-type vmInfo struct {
-	InstanceName instanceName `json:"instanceName,omitempty"`
-	ScalesetHash hash         `json:"scalesetHash,omitempty"`
-}
 
 func hashScaleSets(azuretemplate map[string]interface{}) (map[scalesetName]hash, error) {
 	ssHashes := make(map[scalesetName]hash)
@@ -111,42 +103,4 @@ func (u *simpleUpgrader) initializeUpdateBlob(cs *api.OpenShiftManagedCluster, s
 		}
 	}
 	return u.writeUpdateBlob(vmHashes)
-}
-
-func (u *simpleUpgrader) writeUpdateBlob(b updateblob) error {
-	blob := make([]vmInfo, 0, len(b))
-	for instancename, hash := range b {
-		blob = append(blob, vmInfo{
-			InstanceName: instancename,
-			ScalesetHash: hash,
-		})
-	}
-	data, err := json.Marshal(blob)
-	if err != nil {
-		return err
-	}
-	return u.updateBlob.CreateBlockBlobFromReader(bytes.NewReader(data), nil)
-}
-
-func (u *simpleUpgrader) readUpdateBlob() (updateblob, error) {
-	rc, err := u.updateBlob.Get(nil)
-	if err != nil {
-		return nil, err
-	}
-	defer rc.Close()
-
-	data, err := ioutil.ReadAll(rc)
-	if err != nil {
-		return nil, err
-	}
-
-	var blob []vmInfo
-	if err := json.Unmarshal(data, &blob); err != nil {
-		return nil, err
-	}
-	b := updateblob{}
-	for _, vi := range blob {
-		b[vi.InstanceName] = vi.ScalesetHash
-	}
-	return b, nil
 }
